Fix project#12
Fix lightbox. When clickin in an image after loading the page, lightbox image did not correspond to the image on the clicked image. When refresing the page, this problem did not happen. lightbox.src was removed from html to avoid initial randomness in its src. Also, all the images src were captured in an array after loading. This seems to have solve the problem.